requirements.txt
datasets[stream]
numpy
GitPython
tqdm